Refactor projects page
The projects page was looking unfocused. I added a short blurb to the top,
similar to the one on the about page, and then separated the page into two
main sections: Project Showcase and All Projects. I then removed the minor
projects from the Project Showcase section and added a list with links to
all of my projects to the All Projects section.