Update projects page and add website mockup image
The projects page description was updated to improve readability.
An image with my website being displayed on both a smartphone and
desktop computer was added. Since the original was a little large
I also added a scaled down version of it. The scaled version
looked good so I made it my new project showcase image.