Add video to projects page & change col values
The Neuronmancer screenshot on Projects page was replaced with
an embeded YouTube video. Time will tell if this was a good
decision. Embedding that video introduced new spacing issues so
the col values were updated to make things look better. To be
consistent, the col values were changed on the about page also.
The landing page was left alone though, because it is tailored
to look good on mobile devices and does follow the layout of the
other two pages.